+++ ohsome_quality_analyst/api/api.py
@@ -13,6 +13,7 @@ from fastapi.openapi.docs import (
     get_swagger_ui_oauth2_redirect_html,
 )
 from fastapi.responses import JSONResponse
+from geojson import FeatureCollection
 from starlette.exceptions import HTTPException as StarletteHTTPException
 from starlette.staticfiles import StaticFiles
 
@@ -90,18 +91,6 @@ TAGS_METADATA = [
     {
         "name": "indicator",
         "description": "Request an Indicator",
-        "externalDocs": {
-            "description": "External docs",
-            "url": (
-                "https://github.com/GIScience/ohsome-quality-analyst/blob/"
-                + __version__
-                + "/docs/api.md"
-            ),
-        },
-    },
-    {
-        "name": "report",
-        "description": "Request a Report",
         "externalDocs": {
             "description": "External docs",
             "url": (
@@ -268,9 +257,14 @@ def empty_api_response() -> dict:
 @app.post("/indicators/mapping-saturation/data", include_in_schema=False)
 async def post_indicator_ms(parameters: IndicatorDataRequest) -> CustomJSONResponse:
     """Legacy support for computing the Mapping Saturation indicator for given data."""
-    geojson_object = await oqt.create_indicator(
-        parameters,
+    indicators = await oqt.create_indicator(
         key="mapping-saturation",
+        bpolys=parameters.bpolys,
+        topic=parameters.topic,
+        include_figure=parameters.include_figure,
+    )
+    geojson_object = FeatureCollection(
+        features=[i.as_feature(parameters.include_data) for i in indicators]
     )
     response = empty_api_response()
     response["attribution"]["text"] = get_class_from_key(
@@ -278,8 +272,6 @@ async def post_indicator_ms(parameters: IndicatorDataRequest) -> CustomJSONRespo
         key="mapping-saturation",
     ).attribution()
     # TODO: if accept=JSON no GeoJSON should be created in the first place.
-    #   factor out logic and decision to base/indicator.py and oqt.py
-    #   base/indicator.py should have `as_dict` alongside `as_feature`
     response["results"] = [feature.properties for feature in geojson_object.features]
     return CustomJSONResponse(content=response, media_type=MEDIA_TYPE_JSON)
 
@@ -296,34 +288,27 @@ async def post_indicator_ms(parameters: IndicatorDataRequest) -> CustomJSONRespo
 )
 async def post_indicator(
     request: Request,
-    key: Annotated[
-        IndicatorEnum,
-        Path(
-            title="Indicator Key",
-            example="mapping-saturation",
-        ),
-    ],
+    key: IndicatorEnum,
     parameters: IndicatorRequest,
 ) -> CustomJSONResponse:
-    """Request an Indicator for an AOI defined by OQT or a custom AOI."""
-    if isinstance(parameters, IndicatorRequest):
-        validate_indicator_topic_combination(key.value, parameters.topic_key.value)
-    geojson_object = await oqt.create_indicator(parameters, key=key.value)
-    response = empty_api_response()
-    response["attribution"]["text"] = get_class_from_key(
-        class_type="indicator",
+    """Request an Indicator for a custom AOI."""
+    validate_indicator_topic_combination(key.value, parameters.topic_key.value)
+    indicators = await oqt.create_indicator(
         key=key.value,
-    ).attribution()
-    # TODO: if accept=JSON no GeoJSON should be created in the first place.
-    #   factor out logic and decision to base/indicator.py and oqt.py
-    #   base/indicator.py should have `as_dict` alongside `as_feature`
+        bpolys=parameters.bpolys,
+        topic=get_topic_preset(parameters.topic_key.value),
+        include_figure=parameters.include_figure,
+    )
+
+    response = empty_api_response()
+    response["attribution"]["text"] = indicators[0].attribution()
+
     if request.headers["accept"] == MEDIA_TYPE_JSON:
-        response["results"] = [
-            feature.properties for feature in geojson_object.features
-        ]
+        response["results"] = [i.as_dict(parameters.include_data) for i in indicators]
         return CustomJSONResponse(content=response, media_type=MEDIA_TYPE_JSON)
     elif request.headers["accept"] == MEDIA_TYPE_GEOJSON:
-        response.update(geojson_object)
+        features = [i.as_feature(parameters.include_data) for i in indicators]
+        response.update(FeatureCollection(features))
         return CustomJSONResponse(content=response, media_type=MEDIA_TYPE_GEOJSON)
     else:
         detail = "Content-Type needs to be either {0} or {1}".format(
@@ -334,7 +319,7 @@ async def post_indicator(
         )
 
 
-@app.post("/reports/{key}", tags=["report"])
+@app.post("/reports/{key}", include_in_schema=False)
 async def post_report(
     key: Annotated[
         ReportEnum,
@@ -345,7 +330,6 @@ async def post_report(
     ],
     parameters: ReportRequest,
 ) -> CustomJSONResponse:
-    """Request a Report for an AOI defined by OQT or a custom AOI."""
     geojson_object = await oqt.create_report(parameters, key=key.value)
     response = empty_api_response()
     response["attribution"]["text"] = get_class_from_key(
@@ -366,7 +350,7 @@ async def post_report(
                 k.value: {"label_description": True, "result_description": True}
                 for k in IndicatorEnum
             },
-            "reports": {k.value: {"label_description": True} for k in ReportEnum},
+            # "reports": {k.value: {"label_description": True} for k in ReportEnum},
         }
     },
 )
@@ -379,7 +363,7 @@ async def metadata(project: ProjectEnum = DEFAULT_PROJECT) -> MetadataResponse:
         "quality_dimensions": get_quality_dimensions(),
         "projects": get_project_metadata(),
         "indicators": get_indicator_metadata(project=project),
-        "reports": get_report_metadata(project=project),
+        # "reports": get_report_metadata(project=project),
     }
     return MetadataResponse(result=result)
 
@@ -494,6 +478,7 @@ async def metadata_indicators_by_key(key: IndicatorEnum) -> IndicatorMetadataRes
     response_model_exclude={
         "result": {k.value: {"label_description": True} for k in ReportEnum}
     },
+    include_in_schema=False,
 )
 async def metadata_reports(
     project: ProjectEnum = DEFAULT_PROJECT,
@@ -510,6 +495,7 @@ async def metadata_reports(
     response_model_exclude={
         "result": {k.value: {"label_description": True} for k in ReportEnum}
     },
+    include_in_schema=False,
 )
 async def metadata_reports_by_key(key: ReportEnum) -> ReportMetadataResponse:
     """Get metadata of an indicator by key."""