+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 - api: remove `include_svg`, `include_html` and `flatten` parameter ([#644])
 - api: only accept FeatureCollection as `bpolys` parameter ([#650])
 
+### New Features
+
+- feat(api): add `includeFigure` parameter to `/indicators` endpoint. ([#659])
+
 ### Other Changes
 
 - style: use `ruff` instead `flake8` and `isort` ([#630])
@@ -23,7 +27,9 @@
 - refactor(api): clean-up and rename request models ([#649])
 - refactor(api): move openapi param example to model ([#649])
 - currentness: results with low contributions are now labeled undefined ([#643])
-
+- refactor(api): exclude reports endpoint from schema ([#656])
+- feat(indicator): add as_dict method ([#653])
+- refactor(oqt/api): move geojson creation to api ([#653])
 
 [#589]: https://github.com/GIScience/ohsome-quality-analyst/pull/589
 [#600]: https://github.com/GIScience/ohsome-quality-analyst/pull/600
@@ -38,6 +44,9 @@
 [#647]: https://github.com/GIScience/ohsome-quality-analyst/pull/647
 [#649]: https://github.com/GIScience/ohsome-quality-analyst/pull/649
 [#650]: https://github.com/GIScience/ohsome-quality-analyst/pull/650
+[#653]: https://github.com/GIScience/ohsome-quality-analyst/pull/653
+[#656]: https://github.com/GIScience/ohsome-quality-analyst/pull/656
+[#659]: https://github.com/GIScience/ohsome-quality-analyst/pull/659
 
 ## 0.15.0